refactor(storybook): Renames and cleanup
Renames `StoryTable.story-helpers.tsx` to `StoryTable.tsx` and `SimpleStoryTable.story-helpers.tsx` to `createTableStory.tsx`.

This change improves clarity and consistency in the naming of Storybook utility files. The `createTableStory` function is now more accurately named based on its purpose.
